Level 1 Design + Platforms + Invisible Walls
The level is missing props (like lava I need to add to fill in the gaps in case the player misses a jump and other visual aesthetics' such as more background props in the background etc.) and to finish the script to get moving script for the platforms. Also the platforms already have their box colliders set.